Reject null-returning receivers; hide base fields C#-style; wasm/timer runtime
Compiler
- checker: a member access whose receiver is a call that can return null (or a
  local holding such a result that the body never tests) is an error, with the
  advice to use `?.` or guard the value. This turns the whole class of
  "read at 0x8, minutes later, in unrelated code" faults into a diagnostic.
- binder/irgen: a derived field may repeat a base field's name again. Both keep
  a slot -- the base one stays in the layout prefix, so base methods still touch
  the field they were compiled against -- and a name resolves to the last
  declaration for the static type in hand (C# hiding, including when the two
  differ in type: `class Leaf : Node { int tag; }` over `string tag`). The
  previous behaviour dropped the base slot and made base methods write through
  the wrong offset; the interim diagnostic that rejected it is gone.
- main: `-g` turns on `--check-leaks` and `--arc-guard` by default and
  `--publish` turns both off, so a debug run reports leaks and stale references
  with source locations without anyone remembering a flag. `--no-check-leaks` /
  `--no-arc-guard` opt out.

Runtime
- rt_file.c: file IO split out so wasm32 can provide its own; rt_wasm/rt_timer
  follow, and rt_sync/rt_timer carry the timer-handle fixes.

stdlib / IDE / tests
- Guard the call results the new diagnostic rejects (ZanDb, Firebird SRP,
  WeChat pay, JsonValue, GenDb/GenRoute/GenCommon, Serialize, ChartModel and
  the IDE views/dialogs).
- tests/conformance/field_hiding.zan covers hiding across base/derived methods
  and both static types.

+++ src/runtime/rt_timer.h
@@ -16,8 +16,10 @@ void zan_timer_set_ready_hook(void (*ready)(void *frame, zan_timer_step_t step))
 long long zan_timer_now_ms(void);
 void zan_timer_delay(long long ms, void *frame, zan_timer_step_t step);
 long long zan_timer_next_timeout(void);
-size_t zan_timer_dispatch_due(void);
-size_t zan_timer_pending(void);
+/* These return counts; `long long` (not size_t) because the compiler's IR
+ * declares them with Zan's 64-bit int and wasm32's size_t is 32-bit. */
+long long zan_timer_dispatch_due(void);
+long long zan_timer_pending(void);
 
 long long zan_timer_tick(long long interval, zan_timer_callback_t callback);
 long long zan_timer_after(long long delay, zan_timer_callback_t callback);
